backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
scikit-learn==1.7.1
nltk==3.9.1
numpy==2.3.2
scipy==1.16.1
pymongo==4.14.1
transformers
torch
gensim
networkx
drain3
google-generativeai
python-dotenv
